Add documentation and refactor `_plotting` module
