+++ README.md
@@ -1,23 +1,26 @@
-<div style="width: 50%; margin: auto; margin-bottom: 1.9rem;">
-<picture>
-  <source media="(prefers-color-scheme: dark)" srcset="docs/assets/statsplotly_logo.png">
-  <img alt="Statsplotly Logo" src="docs/assets/statsplotly_logo.png">
-</picture>
+<div align="center" style="margin-bottom: 6vw;">
+
+![Logo Dark](docs/assets/statsplotly-dark-mode-logo.png#gh-dark-mode-only)
 </div>
 
-[![Documentation](https://img.shields.io/badge/Documentation-8A2BE2)](https://parici75.github.io/statsplotly)
+<div align="center">
 
+![Logo Light](docs/assets/statsplotly-light-mode-logo.png#gh-light-mode-only)
+</div>
+
+
+[![Documentation](https://img.shields.io/website?label=docs&url=https://parici75.github.io/statsplotly)](https://parici75.github.io/statsplotly)
 ![PyPI - Python Version](https://img.shields.io/pypi/pyversions/statsplotly)
 [![Black](https://img.shields.io/badge/Code%20style-Black-black)](https://black.readthedocs.io/en/stable/)
 [![linting - Ruff](https://img.shields.io/badge/Linting-Ruff-yellow)](https://docs.astral.sh/ruff/)
 [![mypy](https://img.shields.io/badge/mypy-checked-blue)](https://mypy.readthedocs.io/en/stable/index.html#)
+[![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://pre-commit.com/)
 [![CI](https://github.com/Parici75/statsplotly/actions/workflows/test.yml/badge.svg)](https://github.com/Parici75/statsplotly/actions/workflows/test.yml)
-[![Documentation](https://github.com/Parici75/statsplotly/actions/workflows/sphinx.yml/badge.svg)](https://github.com/Parici75/statsplotly/actions/workflows/sphinx.yml)
-![PyPI - Package Version](https://img.shields.io/pypi/v/statsplotly)
+[![PyPI - Package Version](https://img.shields.io/pypi/v/statsplotly)](https://pypi.org/project/statsplotly/)
 [![GitHub License](https://img.shields.io/github/license/Parici75/statsplotly)](https://github.com/Parici75/statsplotly/blob/main/LICENSE)
 
 ----------------
-[Statsplotly](https://github.com/parici75/statsplotly) is a Python data visualization library based on [Plotly](https://plotly.com/). It provides a high-level interface for drawing interactive statistical data visualization plots with a tidy, declarative API.
+[Statsplotly](https://github.com/parici75/statsplotly) is a Python data visualization library based on [Plotly](https://plotly.com/python/). It provides a high-level interface for drawing interactive statistical data visualization plots with a tidy, declarative API.
 
 
 ![statsplotly-demo](docs/assets/statsplotly-demo.gif)
@@ -25,12 +28,28 @@
 
 ## Philosophy
 
-The inception of this library predates the fantastic [plotly.express](https://plotly.com/python/plotly-express/) API.
+I began to work on this library during my first job outside academia, a few months before the fantastic [plotly.express](https://plotly.com/python/plotly-express/) API was launched. I was looking for a solution to generate visually appealing statistical plots that adapt well to diverse communication mediums, such as web applications, Jupyter Notebooks, presentations, and printed reports.
+
+There are multiple declarative visualization library in the Python ecosystem, each with their own strengths, but none really fit my needs at the time.
+
+Further, I discovered that real-world data analysis, whether in academia or industry, is rarely about creating facet plots. Instead, it revolves around crafting the most effective graphical representation to achieve one of two primary objectives:
+  - **Exploratory** data visualization: the ideal tool should facilitate effortless immersion into the data to uncover valuable insights.
+  - **Explanatory** data visualization: the ideal tool should enable rapid creation of polished, visually appealing plots that effectively support a hypothesis derived from the preceding analysis.
+
+There is no reasons why these two endeavours should require different workflows. On the contrary, working toward both ends using the same tool should increase both analysis and communication efficiency.
+
+I thus set out to design a library combining :
+- a high level interface with graphical representation entities (e.g., plot, barplot, distplot, etc) as entrypoint functions. As intellectually satisfying the [Grammar of Graphics](https://www.tandfonline.com/doi/pdf/10.1198/jcgs.2009.07098) may be, I find this framework to be quite convoluted for expressing pragmatic visualization needs one has when performing real-world data analysis.
+
+- a mainly declarative, yet chainable and customizable by imperative workflow, API. Assembling the appropriate data visualization units across a flexible combination of facets requires keeping control over the different layers of the graphical representations.
+
+- A genuinely interactive plotting interface: In the realm of data visualization beyond scientific publishing, the primary objectives are to explore and convey insights about a phenomenon, business metrics, or a model's performance. This requires a responsive graphic engine allowing for dynamic visualization.
+
 
 `statsplotly` key standout features are :
-- independent processing of color coding scheme, data slicer and plot dimensions.
-- high level interface for [seaborn-like](https://seaborn.pydata.org/tutorial/distributions.html) visualization of data distributions.
-- standard statistical data processing under the hood.
+- an independent processing of color coding scheme, data slicer and plot dimensions.
+- a high level interface for [seaborn-like](https://seaborn.pydata.org/tutorial/distributions.html) visualization of data distributions.
+- statistical data processing under the hood.
 - leveraging of the tidy DataFrame structure for easy styling of plot cues (e.g., marker color, symbol, size, and opacity).
 - sensible cartesian and coloraxis coordinates management across figure subplots.
 
@@ -39,9 +58,7 @@ In summary, `statsplotly` seeks to take advantage of the powerful interactivity 
 
 ## Documentation
 
-Details of the public API can be found in the [documentation](https://parici75.github.io/statsplotly).
-
-Main features of the API are demonstrated in a demo [notebook](https://nbviewer.org/github/parici75/statsplotly/blob/main/docs/notebooks/statsplotly_demo.ipynb).
+Main features and details of the public API can be found in the [documentation](https://parici75.github.io/statsplotly).
 
 
 ## Installation
@@ -67,7 +84,7 @@ make init
 To make the Poetry-managed kernel available for a globally installed Jupyter :
 
 ```bash
-poetry run ipython kernel install --user --name=<KERNEL_NAME>
+poetry run python -m ipykernel install --user --name=<KERNEL_NAME>
 jupyter notebook
 ```
