+++ docs/conf.py
@@ -32,10 +32,12 @@ extensions = [
     "sphinx.ext.doctest",
     "sphinx.ext.autosectionlabel",
     "sphinx.ext.githubpages",
-    "m2r2",
+    "myst_nb",
 ]
 
-source_suffix = [".rst", ".md"]
+# myst and myst_nb
+myst_enable_extensions = ["html_image"]
+jupyter_execute_notebooks = "cache"
 
 # Napoleon settings
 napoleon_google_docstring = True
@@ -53,6 +55,7 @@ napoleon_preprocess_types = False
 napoleon_type_aliases = None
 napoleon_attr_annotations = True
 
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 
@@ -67,6 +70,13 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 # The theme to use for HTML and HTML Help pages. See the documentation for
 # a list of builtin themes.
 html_theme = "furo"
+html_favicon = "assets/favicon.ico"
+
+html_theme_options = {
+    "sidebar_hide_name": True,
+    "light_logo": "statsplotly-light-mode-logo.png",
+    "dark_logo": "statsplotly-dark-mode-logo.png",
+}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
