+++ docs/index.rst
@@ -3,14 +3,31 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
+.. include:: source/getting_started.md
+   :parser: myst_parser.sphinx_
 
-.. mdinclude:: source/getting_started.md
+Features
+---------
+.. toctree::
+   :maxdepth: 2
+   :caption: Features
+
+   notebooks/lineplot.ipynb
+   notebooks/scatterplot.ipynb
+   notebooks/barplot.ipynb
+   notebooks/catplot.ipynb
+   notebooks/heatmap.ipynb
+   notebooks/distplot.ipynb
+   notebooks/jointplot.ipynb
+
+   notebooks/managing_subplots.ipynb
 
 
 API
 ---
 .. toctree::
    :maxdepth: 1
+   :caption: API
 
    source/statsplotly.rst
 
@@ -20,10 +37,11 @@ Indices and tables
 
 * :ref:`genindex`
 * :ref:`modindex`
-* :ref:`search`
 
 
 Changelog
 ---------
 .. toctree::
+   :caption: Changelog
+
    source/changelog_link.rst
