+++ statsplotly/__init__.py
@@ -1,10 +1,15 @@
-from ._plotting.main import barplot, catplot, distplot, heatmap, jointplot, plot
+from ._plotting._barplot import barplot
+from ._plotting._catplot import catplot
+from ._plotting._distplot import distplot
+from ._plotting._heatmap import heatmap
+from ._plotting._jointplot import jointplot
+from ._plotting._plot import plot
 
 __all__ = [
-    "heatmap",
-    "catplot",
     "plot",
     "barplot",
+    "catplot",
+    "heatmap",
     "distplot",
     "jointplot",
 ]
